refactor(backup): extract scrubbed snapshot creation into backup/snapshot.rs
